Refactor socket broadcasting into separate service module
